feat: Add new go build tag no_openziti to reduce build size
closes https://github.com/edgexfoundry/go-mod-bootstrap/issues/789

The usage of OpenZiti packages to support zero trust feature significantly increases
the build size. For example, core-metadata increases from 14MB to 21MB, core-command
increases from 9.8MB to 17MB, device-virtual increases from 18MB to 31MB, and
app-service-configurable increases from 22Mb to 34MB. As many edge user scenarios
require to deploy EdgeX services on resource-constrained devices without security,
allowing that the services can be built without OpenZiti packages and ZeroTrust
features can be helpful to user cases which don't need zero trust feature.

This commit refactors the codes importing openziti packages with //go:build !no_openziti
directive and creates para codes that don't use openziti packages with
//go:build no_openziti directive, so users can simply build services by specifying
no_openziti tag.

Also remove vestigal zc variables and ZitiContext struct per discussion
with https://github.com/dovholuknf in
https://github.com/edgexfoundry/go-mod-bootstrap/pull/659#discussion_r1834363537
