+++ bootstrap/handlers/httpserver.go
@@ -1,6 +1,6 @@
 /*******************************************************************************
  * Copyright 2019 Dell Inc.
- * Copyright 2021-2023 IOTech Ltd
+ * Copyright 2021-2024 IOTech Ltd
  * Copyright 2023 Intel Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
@@ -19,13 +19,10 @@ package handlers
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"net"
 	"net/http"
 	"os"
 	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -33,17 +30,13 @@ import (
 	"github.com/edgexfoundry/go-mod-core-contracts/v4/common"
 	commonDTO "github.com/edgexfoundry/go-mod-core-contracts/v4/dtos/common"
 
-	"github.com/edgexfoundry/go-mod-bootstrap/v4/bootstrap/config"
 	"github.com/edgexfoundry/go-mod-bootstrap/v4/bootstrap/container"
 	"github.com/edgexfoundry/go-mod-bootstrap/v4/bootstrap/startup"
+	"github.com/edgexfoundry/go-mod-bootstrap/v4/bootstrap/zerotrust"
 	"github.com/edgexfoundry/go-mod-bootstrap/v4/di"
 
-	"github.com/edgexfoundry/go-mod-bootstrap/v4/bootstrap/zerotrust"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
-	edge_apis "github.com/openziti/sdk-golang/edge-apis"
-	"github.com/openziti/sdk-golang/ziti"
-	"github.com/openziti/sdk-golang/ziti/edge"
 )
 
 // HttpServer contains references to dependencies required by the http server implementation.
@@ -53,11 +46,6 @@ type HttpServer struct {
 	doListenAndServe bool
 	serverKey        string
 }
-
-type ZitiContext struct {
-	c *ziti.Context
-}
-type OpenZitiIdentityKey struct{}
 
 // NewHttpServer is a factory method that returns an initialized HttpServer receiver struct.
 func NewHttpServer(router *echo.Echo, doListenAndServe bool, serviceKey string) *HttpServer {
@@ -139,8 +127,6 @@ func (b *HttpServer) BootstrapHandler(
 		Timeout: timeout,
 	}))
 
-	zc := &ZitiContext{}
-
 	b.router.Use(RequestLimitMiddleware(bootstrapConfig.Service.MaxRequestSize, lc))
 
 	b.router.Use(ProcessCORS(bootstrapConfig.Service.CORSConfiguration))
@@ -153,7 +139,6 @@ func (b *HttpServer) BootstrapHandler(
 		Handler:           b.router,
 		ReadHeaderTimeout: 5 * time.Second, // G112: A configured ReadHeaderTimeout in the http.Server averts a potential Slowloris Attack
 	}
-	server.ConnContext = mutator
 
 	wg.Add(1)
 	go func() {
@@ -174,76 +159,7 @@ func (b *HttpServer) BootstrapHandler(
 		}()
 
 		b.isRunning = true
-		listenMode := strings.ToLower(bootstrapConfig.Service.SecurityOptions[config.SecurityModeKey])
-		switch listenMode {
-		case zerotrust.ZeroTrustMode:
-			secretProvider := container.SecretProviderExtFrom(dic.Get)
-			if secretProvider == nil {
-				err = errors.New("secret provider is nil. cannot proceed with zero trust configuration")
-				break
-			}
-			secretProvider.EnableZeroTrust() //mark the secret provider as zero trust enabled
-			var zitiCtx ziti.Context
-			var ctxErr error
-			jwt, jwtErr := secretProvider.GetSelfJWT()
-			if jwtErr != nil {
-				lc.Errorf("could not load jwt: %v", jwtErr)
-				err = jwtErr
-				break
-			}
-			ozUrl := bootstrapConfig.Service.SecurityOptions["OpenZitiController"]
-			if !strings.Contains(ozUrl, "://") {
-				ozUrl = "https://" + ozUrl
-			}
-			caPool, caErr := ziti.GetControllerWellKnownCaPool(ozUrl)
-			if caErr != nil {
-				err = caErr
-				break
-			}
-
-			credentials := edge_apis.NewJwtCredentials(jwt)
-			credentials.CaPool = caPool
-
-			cfg := &ziti.Config{
-				ZtAPI:       ozUrl + "/edge/client/v1",
-				Credentials: credentials,
-			}
-			cfg.ConfigTypes = append(cfg.ConfigTypes, "all")
-
-			zitiCtx, ctxErr = ziti.NewContext(cfg)
-			if ctxErr != nil {
-				err = ctxErr
-				break
-			}
-
-			ozServiceName := zerotrust.OpenZitiServicePrefix + b.serverKey
-			lc.Infof("Using OpenZiti service name: %s", ozServiceName)
-			for t.HasNotElapsed() {
-				ln, listenErr := zitiCtx.Listen(ozServiceName)
-				if listenErr != nil {
-					err = fmt.Errorf("could not bind service %s: %s", ozServiceName, listenErr.Error())
-					t.SleepForInterval()
-				} else {
-					zc.c = &zitiCtx
-					lc.Infof("listening on overlay network. ListenMode '%s' at %s", listenMode, addr)
-					err = server.Serve(ln)
-					break
-				}
-			}
-			if !t.HasNotElapsed() {
-				lc.Error("could not listen on the OpenZiti overlay network. timeout reached")
-			}
-		case "http":
-			fallthrough
-		default:
-			lc.Infof("listening on underlay network. ListenMode '%s' at %s", listenMode, addr)
-			ln, listenErr := net.Listen("tcp", addr)
-			if listenErr != nil {
-				err = listenErr
-				break
-			}
-			err = server.Serve(ln)
-		}
+		err = zerotrust.ListenOnMode(bootstrapConfig, b.serverKey, addr, t, server, dic)
 
 		// "Server closed" error occurs when Shutdown above is called in the Done processing, so it can be ignored
 		if err != nil && err != http.ErrServerClosed {
@@ -291,12 +207,4 @@ func RequestLimitMiddleware(sizeLimit int64, lc logger.LoggingClient) echo.Middl
 			return next(c)
 		}
 	}
-}
-
-func mutator(srcCtx context.Context, c net.Conn) context.Context {
-	if zitiConn, ok := c.(edge.Conn); ok {
-		return context.WithValue(srcCtx, OpenZitiIdentityKey{}, zitiConn)
-	}
-
-	return srcCtx
 }
